Backport paths reorganization from v1.0
