+++ lib/vertebrae/authorization.rb
