+++ lib/vertebrae/base.rb
