+++ lib/vertebrae/configuration.rb
