+++ lib/vertebrae/constants.rb
