+++ lib/vertebrae/model.rb
