+++ lib/vertebrae/response/raise_error.rb
