+++ lib/vertebrae/response_error.rb
