+++ lib/vertebrae/version.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module Vertebrae
+  VERSION = '1.0.0'
+end
