+++ vertebrae.gemspec
@@ -1,8 +1,8 @@
-# Generated by juwelier
-# DO NOT EDIT THIS FILE DIRECTLY
-# Instead, edit Juwelier::Tasks in Rakefile, and run 'rake gemspec'
-# -*- encoding: utf-8 -*-
-# stub: vertebrae 0.6.2 ruby lib
+# frozen_string_literal: true
+
+lib = File.expand_path('lib', __dir__)
+$LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
+require 'version'
 
 Gem::Specification.new do |s|
   s.name = "vertebrae".freeze
@@ -54,7 +54,6 @@ Gem::Specification.new do |s|
   ]
   s.homepage = "http://github.com/controlshift/vertebrae".freeze
   s.licenses = ["MIT".freeze]
-  s.rubygems_version = "3.1.6".freeze
   s.summary = "API Client Infrastructure".freeze
 
   if s.respond_to? :specification_version then
